Fix results list being cut-off + overall style improvements
This commits fixes the results list being cut-off, basically due to
the `justify-content: center` which is not really necessary since
no vertical alignment is needed in a top-to-bottom list.

This also removes the `search-template-container` div which is also
not necessary as all styles can be handled from the other `search-container`
wrapper.